requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
SQLAlchemy==2.0.36
greenlet==3.1.1
feedparser==6.0.11
sentence-transformers==3.2.1
scikit-learn==1.5.2
numpy==2.1.2
python-dotenv==1.0.1
